Unicorn Backpatching to avoid repeated mmio exceptions
Currently this still has issues: It's unable to execute instruction
blocks smaller than 5 bytes and is still pretty slow but it's a decent proof-of-concept.

Slowness is caused by multiple reasons:
1. Unicorn isn't optimised for single instructions, but entire code blocks.
Fixing this requires a MASSIVE rework of our execution model and patching
mechanism, so is out of scope for now.
2. NV2A is still missing a surface cache based on read/write watches
3. Instructions <5 bytes are used heavily, and these still trigger
exceptions. This can be solved, but currently Unicorn crashes often when asked
to execute blocks larger than 1 instruction.