Begin refactoring code into a better structure:
Xbox class is owner of all content relating to the Xbox console:
Currently this is only hardware devices, but will be extended to be a
central point of ownership for Xbe data, Memory, Hardware and everything
else about the emulated Xbox machine.

All other aspects of the emulator will be able to query the emulated
system state, or access memory, etc by querying/calling into the Xbox
class.

(This transition is incomplete, but still compiles/works)