Began building IX86CPU interface class.
This will be the interface class in which X86 CPU emulation implementations will implement
An interface class is being used so we can support multiple emulation backends.

For example:
Interpreter
Recompiler
Unicorn Engine
Virtualizaton (HAXM, KVM, Windows HV Platform, etc).

Chances are we will only support one or two of these options, but the
possibility exists to implement more in the future based on this interface