Enable LLE USB
Currently, this only works for attached game-controllers.

TODO: Update this to support fully configurable controllers/input
devices.

TODO: Make it possible in the GUI to select which Xbox device to emulate on each
controller port
TODO: Make it possible in the GUI to configure which PC input device and
buttons are mapped to the Xbox Input Device
TODO: (For Chihiro) extend to add support for Chihiro Baseboard